requirements.txt
gradio>=5.20.0
numpy>=1.20.0
ffmpeg-python
pillow>=9.0.0
julia>=0.5.7 